// src/app/shared/models/DoctorRevenue.ts
export interface DoctorRevenue {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: string;
    checkPrice: number;
    ourCheckupTotalRevenu: number;
    drCheckupTotalRevenu: number;
    gender_En: string;
    gender_Ar: string;
    country_En: string;
    country_Ar: string;
    governorate_En: string;
    governorate_Ar: string;
    district_En: string;
    district_Ar: string;
}